requirements-app.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.8.0
requests>=2.31.0
